+++ src/main/java/dev/nincodedo/bluemapbanners/BlueMapBanners.java
@@ -4,6 +4,11 @@ import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.tree.LiteralCommandNode;
 import de.bluecolored.bluemap.api.BlueMapAPI;
+import dev.nincodedo.bluemapbanners.manager.ConfigManager;
+import dev.nincodedo.bluemapbanners.manager.MarkerManager;
+import dev.nincodedo.bluemapbanners.util.Compatibility;
+import dev.nincodedo.bluemapbanners.util.MapIcons;
+import dev.nincodedo.bluemapbanners.util.Metrics;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
